+++ pkg/vm/internal/data/sorter_multi.go
@@ -44,7 +44,7 @@ func (s *MultiSorter) Set(ctx context.Context, key, value runtime.Value) error {
 }
 
 func (s *MultiSorter) sort(ctx context.Context) error {
-	return runtime.SortListWith(ctx, s.Value, func(first, second runtime.Value) int64 {
+	return runtime.SortListWith(ctx, s.Value, func(first, second runtime.Value) int {
 		firstKV := first.(*KV)
 		secondKV := second.(*KV)
